-added enums for KeyWords, Default Files -moved schema related files in schema package -moved SchemaToRender to his own file -added TemplateRenderer
