+++ json_schema_for_humans/const.py
@@ -7,30 +7,6 @@ TYPE_NUMBER = "number"
 TYPE_OBJECT = "object"
 TYPE_STRING = "string"
 
-KW_REQUIRED = "required"
-KW_TITLE = "title"
-KW_CONTAINS = "contains"
-KW_ITEMS = "items"
-KW_UNIQUE_ITEMS = "uniqueItems"
-KW_ADDITIONAL_ITEMS = "additionalItems"
-KW_MAX_ITEMS = "maxItems"
-KW_MIN_ITEMS = "minItems"
-KW_MAX_LENGTH = "maxLength"
-KW_MIN_LENGTH = "minLength"
-KW_PATTERN = "pattern"
-KW_CONST = "const"
-KW_ENUM = "enum"
-KW_ELSE = "else"
-KW_THEN = "then"
-KW_IF = "if"
-KW_NOT = "not"
-KW_ONE_OF = "oneOf"
-KW_ANY_OF = "anyOf"
-KW_ALL_OF = "allOf"
-KW_PROPERTIES = "properties"
-KW_PATTERN_PROPERTIES = "patternProperties"
-KW_ADDITIONAL_PROPERTIES = "additionalProperties"
-
 DESCRIPTION = "description"
 DEFAULT = "default"
 EXAMPLES = "examples"
