+++ json_schema_for_humans/default_file.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class DefaultFile(Enum):
+    CSS_FILE_NAME = "schema_doc.css"
+    JS_FILE_NAME = "schema_doc.min.js"
+    TEMPLATE_FILE_NAME = "base.html"
