+++ json_schema_for_humans/language_types.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class LanguageTypes(Enum):
+    js = "js"
+    md = "md"
+    md_nested = "md_nested"
+    html = "html"
