+++ json_schema_for_humans/schema/schema_node.py
@@ -1,9 +1,8 @@
 import copy
 from typing import Any, Dict, Iterable, Iterator, List, Optional, Set, Union, cast
 
-import jinja2
-
 from json_schema_for_humans import const
+from json_schema_for_humans.schema.keyword import KeyWord
 from json_schema_for_humans.templating_utils import get_type_name
 from json_schema_for_humans.generation_configuration import GenerationConfiguration
 
@@ -136,7 +135,7 @@ class SchemaNode:
 
     @property
     def is_additional_properties(self) -> bool:
-        return self.parent_key == const.KW_ADDITIONAL_PROPERTIES
+        return self.parent_key == KeyWord.ADDITIONAL_PROPERTIES.value
 
     @property
     def is_a_property_node(self) -> bool:
@@ -190,7 +189,7 @@ class SchemaNode:
     def path_to_property(self) -> str:
         """Human-readable representation of the path from the root of the schema to this node"""
         path_without_properties = [
-            p for p in self.path_to_element if p not in [const.KW_PROPERTIES, const.KW_PATTERN_PROPERTIES]
+            p for p in self.path_to_element if p not in [KeyWord.PROPERTIES.value, KeyWord.PATTERN_PROPERTIES.value]
         ]
         return " -> ".join([p if isinstance(p, str) else f"Item {p}" for p in path_without_properties])
 
@@ -273,11 +272,11 @@ class SchemaNode:
 
         # return self._refers_to_merged
 
-    def get_keyword(self, keyword: str) -> Optional["SchemaNode"]:
+    def get_keyword(self, keyword: KeyWord) -> Optional["SchemaNode"]:
         """Get the value of a keyword if present and it is not a property (to avoid conflicts with properties being
         named like a keyword, e.g. a property named "if")
         """
-        possible_keyword = self.keywords.get(keyword)
+        possible_keyword = self.keywords.get(keyword.value)
         if possible_keyword and isinstance(possible_keyword, SchemaNode) and not possible_keyword.is_property:
             return possible_keyword
 
@@ -285,19 +284,19 @@ class SchemaNode:
 
     @property
     def kw_all_of(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_ALL_OF)
+        return self.get_keyword(KeyWord.ALL_OF)
 
     @property
     def kw_any_of(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_ANY_OF)
+        return self.get_keyword(KeyWord.ANY_OF)
 
     @property
     def kw_one_of(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_ONE_OF)
+        return self.get_keyword(KeyWord.ONE_OF)
 
     @property
     def kw_not(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_NOT)
+        return self.get_keyword(KeyWord.NOT)
 
     @property
     def has_conditional(self) -> bool:
@@ -305,52 +304,52 @@ class SchemaNode:
 
     @property
     def kw_if(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_IF)
+        return self.get_keyword(KeyWord.IF)
 
     @property
     def kw_then(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_THEN)
+        return self.get_keyword(KeyWord.THEN)
 
     @property
     def kw_else(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_ELSE)
+        return self.get_keyword(KeyWord.ELSE)
 
     @property
     def kw_enum(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_ENUM)
+        return self.get_keyword(KeyWord.ENUM)
 
     @property
     def kw_const(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_CONST)
+        return self.get_keyword(KeyWord.CONST)
 
     @property
     def kw_pattern(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_PATTERN)
+        return self.get_keyword(KeyWord.PATTERN)
 
     @property
     def kw_properties(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_PROPERTIES)
+        return self.get_keyword(KeyWord.PROPERTIES)
 
     @property
     def kw_pattern_properties(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_PATTERN_PROPERTIES)
+        return self.get_keyword(KeyWord.PATTERN_PROPERTIES)
 
     @property
     def kw_additional_properties(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_ADDITIONAL_PROPERTIES)
+        return self.get_keyword(KeyWord.ADDITIONAL_PROPERTIES)
 
     @property
     def kw_min_length(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_MIN_LENGTH)
+        return self.get_keyword(KeyWord.MIN_LENGTH)
 
     @property
     def kw_max_length(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_MAX_LENGTH)
+        return self.get_keyword(KeyWord.MAX_LENGTH)
 
     @property
     def kw_items(self) -> Optional[List["SchemaNode"]]:
         """items can be either an object either a list of object"""
-        items = self.get_keyword(const.KW_ITEMS)
+        items = self.get_keyword(KeyWord.ITEMS)
         if not items:
             return None
 
@@ -361,31 +360,31 @@ class SchemaNode:
 
     @property
     def kw_min_items(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_MIN_ITEMS)
+        return self.get_keyword(KeyWord.MIN_ITEMS)
 
     @property
     def kw_max_items(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_MAX_ITEMS)
+        return self.get_keyword(KeyWord.MAX_ITEMS)
 
     @property
     def kw_unique_items(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_UNIQUE_ITEMS)
+        return self.get_keyword(KeyWord.UNIQUE_ITEMS)
 
     @property
     def kw_additional_items(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_ADDITIONAL_ITEMS)
+        return self.get_keyword(KeyWord.ADDITIONAL_ITEMS)
 
     @property
     def kw_contains(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_CONTAINS)
+        return self.get_keyword(KeyWord.CONTAINS)
 
     @property
     def kw_required(self) -> Optional["SchemaNode"]:
-        return self.get_keyword(const.KW_REQUIRED)
+        return self.get_keyword(KeyWord.REQUIRED)
 
     @property
     def title(self) -> Optional[str]:
-        title_kw = self.get_keyword(const.KW_TITLE)
+        title_kw = self.get_keyword(KeyWord.TITLE)
         if not title_kw:
             return None
         title = title_kw.literal
