+++ tests/intermediate_test.py
@@ -1,15 +1,14 @@
-from json_schema_for_humans.intermediate_representation import build_intermediate_representation
-from json_schema_for_humans.generation_configuration import GenerationConfiguration
+from json_schema_for_humans.schema.intermediate_representation import build_intermediate_representation
 from tests.test_utils import get_test_case_path
 
 
 def test_basic() -> None:
-    intermediate = build_intermediate_representation(get_test_case_path("basic"), GenerationConfiguration())
+    intermediate = build_intermediate_representation(get_test_case_path("basic"))
 
     assert intermediate
 
 
 def test_references() -> None:
-    intermediate = build_intermediate_representation(get_test_case_path("references"), GenerationConfiguration())
+    intermediate = build_intermediate_representation(get_test_case_path("references"))
 
     assert intermediate
