+++ tests/test_md_generate.py
@@ -4,7 +4,7 @@ import pytest
 from dataclasses import dataclass
 from pathlib import Path
 
-from json_schema_for_humans.generation_configuration import GenerationConfiguration
+from json_schema_for_humans.generation_configuration import GenerationConfiguration, LanguageTypes
 from tests.md_utils_asserts import MdUtilsAsserts
 
 current_dir = os.path.abspath(os.path.dirname(__file__))
@@ -12,10 +12,10 @@ parent_dir = os.path.abspath(os.path.dirname(current_dir))
 examples_dir = os.path.join(parent_dir, "docs", "examples")
 
 config_badge = GenerationConfiguration(
-    template_name="md", template_md_options={"badge_as_image": True}, deprecated_from_description=True
+    template_name=LanguageTypes.md, template_md_options={"badge_as_image": True}, deprecated_from_description=True
 )
 config_no_badge = GenerationConfiguration(
-    template_name="md", template_md_options={"badge_as_image": False}, deprecated_from_description=True
+    template_name=LanguageTypes.md, template_md_options={"badge_as_image": False}, deprecated_from_description=True
 )
