+++ tests/test_utils.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from bs4 import BeautifulSoup
 
 from json_schema_for_humans.generate import generate_from_schema
-from json_schema_for_humans.generation_configuration import GenerationConfiguration
+from json_schema_for_humans.generation_configuration import GenerationConfiguration, DefaultFile
 
 current_dir = os.path.dirname(os.path.abspath(__file__))
 parent_dir = os.path.dirname(current_dir)
@@ -12,18 +12,16 @@ parent_dir = os.path.dirname(current_dir)
 
 def assert_css_and_js_not_copied(path: Path) -> None:
     """Assert that the standard CCS and JS files were not copied over"""
-    assert not (path / "schema_doc.css").exists()
+    assert not (path / DefaultFile.CSS_FILE_NAME.value).exists()
 
-    assert not (path / "schema_doc.min.js").exists()
+    assert not (path / DefaultFile.JS_FILE_NAME.value).exists()
 
 
-def assert_css_and_js_copied(
-    path: Path, css_file_name: str = "schema_doc.css", js_file_name: str = "schema_doc.min.js"
-) -> None:
+def assert_css_and_js_copied(path: Path) -> None:
     """Assert that the standard CCS and JS files were copied over"""
-    assert (path / css_file_name).exists()
+    assert (path / DefaultFile.CSS_FILE_NAME.value).exists()
 
-    assert (path / js_file_name).exists()
+    assert (path / DefaultFile.JS_FILE_NAME.value).exists()
 
 
 def get_test_cases_path() -> str:
